Split OffsetCommitHandler into three classes to better decouple them.
